move FailureSafeParser from catalyst to sql core
